requirements.txt
psycopg2==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
